pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.11,<3.14"
dependencies = [
    "pillow>=12.1.1",
    "python-telegram-bot>=22.6",
    "qrcode>=8.2",
    "flask>=3.1.0",
    "psycopg2-binary>=2.9",
]
